🔧 Fix CI/CD JSON formatting + Add Quality Enhancement Pipeline
- Fixed JSON format error in swift-cleanup workflow
- Added post-cleanup quality enhancement automation
- Prepared next optimization phase: 78 async functions, 13 NeuroMesh files
- Ready for: Component consolidation, Performance tuning, Apple Silicon

+++ NovaMind/Core/Color+Theme.swift
@@ -17,7 +17,7 @@ extension Color {
 
     // MARK: - Computed Colors (for dynamic behavior)
     static var interactiveGlow: Color {
-        glow.opacity(0.8)
+        glow.opacity(0.8 as Double)
     }
 
     static var panelBackground: Color {
